DroneController:  * Area Coverage: Added variables to be able to control search line width and height difference between searching drones (they do have default values, so SystemController can set them only if it wants to).  * Bridge: Added default value for max comm distance for Bridge, so that SystemController may not set it if it doesn't want to.  * Bridge: Fixed bug in Bridge algorithm which was not expanding a statement (change in Madara removed auto-expansion).  * Platforms: renamed all platform interface functions to have the "platfrom_" prefix. This makes it much easier to spot platform-specific funtions in the Sensor and Movement modules.  * Area_Coverage: minor code standards cleanup.  * AR2 Drone Platform: minor cleanup of unused variables, and reordering of functions according to function.  * Linux Platform: brought up to date with missing new interface functions.
Simulation:
 * Added vbs script to start 4 drones on console windows evenly distributed over the screen for easier debugging of their behavior.
 * Initialized thermal buffer in VRep Platform to all-zeros, to ensure there is no garbage being read if there is no thermal reading from VRep.

+++ DroneController/packages/madara_client/area_coverage/InsideOutAreaCoverage.cpp
@@ -23,7 +23,7 @@ using namespace SMASH::Utilities;
 using namespace SMASH::AreaCoverage;
 
 // Constructor
-InsideOutAreaCoverage::InsideOutAreaCoverage(float delta, direction_t heading,
+InsideOutAreaCoverage::InsideOutAreaCoverage(double delta, direction_t heading,
     bool clockwise) : AreaCoverage(), m_delta(delta), m_iteration(2), m_clockwise(clockwise),
     m_heading(heading), m_nsMultiplier(1.0), m_ewMultiplier(1.0) {}
